Rename Error component to ErrorDisplay (to avoid shadowing Error)
